Move tests to correct namespace
