+++ tests/Jellyfin.NfoMetadata.Tests/Parsers/MovieNfoParserTests.cs
@@ -16,7 +16,7 @@ using Microsoft.Extensions.Logging.Abstractions;
 using Moq;
 using Xunit;
 
-namespace Jellyfin.XbmcMetadata.Parsers.Tests
+namespace Jellyfin.NfoMetadata.Tests.Parsers
 {
     public class MovieNfoParserTests
     {
