+++ tests/Jellyfin.NfoMetadata.Tests/Parsers/MusicArtistNfoParserTests.cs
@@ -17,7 +17,7 @@ using Microsoft.Extensions.Logging.Abstractions;
 using Moq;
 using Xunit;
 
-namespace Jellyfin.XbmcMetadata.Parsers.Tests
+namespace Jellyfin.NfoMetadata.Tests.Parsers
 {
     public class MusicArtistNfoParserTests
     {
