+++ tests/Jellyfin.NfoMetadata.Tests/Test Data/Justice League.nfo
