+++ tests/Jellyfin.NfoMetadata.Tests/Test Data/U2.nfo
